remove old gateway url, rename views_beta to views
