+++ gateway/tests/_test_data_mesh.py
@@ -1,3 +1,4 @@
+# TODO: remove, this is for old version of the gateway (_views.py)
 import uuid
 
 import factories
