+++ gateway/tests/test_views.py
